refactor: move IntegrationTestSetup
commit-id:541ff2dd
