+++ crates/starknet_integration_tests/src/integration_test_setup.rs
@@ -2,11 +2,8 @@ use std::net::SocketAddr;
 use std::path::PathBuf;
 
 use blockifier::context::ChainInfo;
-use mempool_test_utils::starknet_api_test_utils::{
-    AccountTransactionGenerator,
-    MultiAccountTransactionGenerator,
-};
-use papyrus_storage::{StorageConfig, StorageReader};
+use mempool_test_utils::starknet_api_test_utils::AccountTransactionGenerator;
+use papyrus_storage::StorageConfig;
 use starknet_api::rpc_transaction::RpcTransaction;
 use starknet_api::transaction::TransactionHash;
 use starknet_consensus_manager::config::ConsensusManagerConfig;
@@ -17,103 +14,12 @@ use starknet_monitoring_endpoint::config::MonitoringEndpointConfig;
 use starknet_monitoring_endpoint::test_utils::MonitoringClient;
 use starknet_sequencer_infra::test_utils::AvailablePorts;
 use starknet_sequencer_node::config::component_config::ComponentConfig;
-use starknet_sequencer_node::test_utils::node_runner::spawn_run_node;
 use tempfile::{tempdir, TempDir};
-use tokio::task::JoinHandle;
-use tracing::{info, instrument};
+use tracing::instrument;
 
 use crate::config_utils::dump_config_file_changes;
 use crate::state_reader::StorageTestSetup;
-use crate::utils::{
-    create_chain_info,
-    create_consensus_manager_configs_and_channels,
-    create_mempool_p2p_configs,
-    create_node_config,
-};
-
-pub struct IntegrationTestSetup {
-    pub sequencers: Vec<IntegrationSequencerSetup>,
-    pub sequencer_run_handles: Vec<JoinHandle<()>>,
-}
-
-impl IntegrationTestSetup {
-    pub async fn run(
-        tx_generator: &MultiAccountTransactionGenerator,
-        mut available_ports: AvailablePorts,
-        component_configs: Vec<Vec<ComponentConfig>>,
-    ) -> Self {
-        let chain_info = create_chain_info();
-        let accounts = tx_generator.accounts();
-        let n_distributed_sequencers =
-            component_configs.iter().map(|inner_vec| inner_vec.len()).sum();
-
-        let (mut consensus_manager_configs, _) = create_consensus_manager_configs_and_channels(
-            n_distributed_sequencers,
-            &mut available_ports,
-        );
-
-        let ports = available_ports.get_next_ports(n_distributed_sequencers);
-        let mut mempool_p2p_configs =
-            create_mempool_p2p_configs(chain_info.chain_id.clone(), ports);
-
-        let mut sequencers = vec![];
-        for (sequencer_id, node_composition) in component_configs.iter().enumerate() {
-            for component_config in node_composition {
-                // Declare one consensus_manager_config and one mempool_p2p_config for each node
-                // composition
-                let consensus_manager_config = consensus_manager_configs.remove(0);
-                let mempool_p2p_config = mempool_p2p_configs.remove(0);
-                let sequencer = IntegrationSequencerSetup::new(
-                    accounts.to_vec(),
-                    sequencer_id,
-                    chain_info.clone(),
-                    consensus_manager_config,
-                    mempool_p2p_config,
-                    &mut available_ports,
-                    component_config.clone(),
-                )
-                .await;
-                sequencers.push(sequencer);
-            }
-        }
-
-        info!("Running sequencers.");
-        let sequencer_run_handles = sequencers
-            .iter()
-            .map(|sequencer| spawn_run_node(sequencer.node_config_path.clone()))
-            .collect::<Vec<_>>();
-
-        Self { sequencers, sequencer_run_handles }
-    }
-
-    pub async fn await_alive(&self, interval: u64, max_attempts: usize) {
-        for (sequencer_index, sequencer) in self.sequencers.iter().enumerate() {
-            sequencer
-                .monitoring_client
-                .await_alive(interval, max_attempts)
-                .await
-                .unwrap_or_else(|_| panic!("Node {} should be alive.", sequencer_index));
-        }
-    }
-
-    pub async fn send_rpc_tx_fn(&self, rpc_tx: RpcTransaction) -> TransactionHash {
-        self.sequencers[0].assert_add_tx_success(rpc_tx).await
-    }
-
-    pub fn batcher_storage_reader(&self) -> StorageReader {
-        let (batcher_storage_reader, _) =
-            papyrus_storage::open_storage(self.sequencers[0].batcher_storage_config.clone())
-                .expect("Failed to open batcher's storage");
-        batcher_storage_reader
-    }
-
-    pub fn shutdown_nodes(&self) {
-        self.sequencer_run_handles.iter().for_each(|handle| {
-            assert!(!handle.is_finished(), "Node should still be running.");
-            handle.abort()
-        });
-    }
-}
+use crate::utils::create_node_config;
 
 pub struct IntegrationSequencerSetup {
     /// Used to differentiate between different sequencer nodes.
